+++ tests/test_pyquery.py
@@ -3,141 +3,57 @@
 # Copyright (C) 2008 - Olivier Lauzanne <olauzanne@gmail.com>
 #
 # Distributed under the BSD license, see LICENSE.txt
-from webob import Request, Response, exc
-from lxml import etree
-import unittest
-import doctest
-import socket
-import sys
 import os
-
-PY3k = sys.version_info >= (3,)
-
-if PY3k:
-    from io import StringIO
-    import pyquery
-    from pyquery.pyquery import PyQuery as pq
-    from pyquery.ajax import PyQuery as pqa
-    from http.client import HTTPConnection
-    text_type = str
-
-    def u(value, encoding):
-        return str(value)
-
-    def b(value):
-        return value.encode('utf-8')
-else:
-    from cStringIO import StringIO  # NOQA
-    import pyquery  # NOQA
-    from httplib import HTTPConnection  # NOQA
-    from pyquery import PyQuery as pq  # NOQA
-    from ajax import PyQuery as pqa  # NOQA
-    text_type = unicode
-
-    def u(value, encoding):  # NOQA
-        return unicode(value, encoding)
-
-    def b(value):  # NOQA
-        return str(value)
+import sys
+sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
+from lxml import etree
+from pyquery.pyquery import PyQuery as pq
+from pyquery.ajax import PyQuery as pqa
+from webtest import http
+from webtest.debugapp import debug_app
+from .apps import application
+from .apps import secure_application
+from .compat import PY3k
+from .compat import u
+from .compat import b
+from .compat import text_type
+from .compat import TestCase
 
 
 def not_py3k(func):
     if not PY3k:
         return func
 
-socket.setdefaulttimeout(5)
-
 try:
-    conn = HTTPConnection("packages.python.org:80")
-    conn.request("GET", "/pyquery/")
-    response = conn.getresponse()
-except (socket.timeout, socket.error):
-    GOT_NET = False
-else:
-    GOT_NET = True
-
-try:
-    import requests # NOQA
+    import requests  # NOQA
     HAS_REQUEST = True
 except ImportError:
     HAS_REQUEST = False
 
 
-def with_net(func):
-    if GOT_NET:
-        return func
-
-dirname = os.path.dirname(os.path.abspath(pyquery.__file__))
+dirname = os.path.dirname(os.path.abspath(__file__))
 docs = os.path.join(os.path.dirname(dirname), 'docs')
 path_to_html_file = os.path.join(dirname, 'test.html')
 
 
-def input_app(environ, start_response):
-    resp = Response()
-    req = Request(environ)
-    if req.path_info == '/':
-        resp.body = '<input name="youyou" type="text" value="" />'
-    elif req.path_info == '/submit':
-        resp.body = '<input type="submit" value="OK" />'
-    else:
-        resp.body = ''
-    return resp(environ, start_response)
-
-
-class TestReadme(doctest.DocFileCase):
-    path = os.path.join(dirname, '..', 'README.rst')
-
-    def __init__(self, *args, **kwargs):
-        parser = doctest.DocTestParser()
-        fd = open(self.path)
-        doc = fd.read()
-        fd.close()
-        test = parser.get_doctest(doc, globals(), '', self.path, 0)
-        doctest.DocFileCase.__init__(self, test, optionflags=doctest.ELLIPSIS)
-
-    def setUp(self):
-        test = self._dt_test
-        test.globs.update(globals())
-
-for filename in os.listdir(docs):
-    if filename.endswith('.txt'):
-        if not GOT_NET and filename in ('ajax.txt', 'tips.txt', 'scrap.txt'):
-            continue
-        if not HAS_REQUEST and filename in ('scrap.txt',):
-            continue
-        if PY3k and filename in ('ajax.txt'):
-            continue
-        klass_name = 'Test%s' % filename.replace('.txt', '').title()
-        path = os.path.join(docs, filename)
-        exec('%s = type("%s", (TestReadme,), dict(path=path))' % (
-                                                       klass_name, klass_name))
-
-
-class TestTests(doctest.DocFileCase):
-    path = os.path.join(dirname, 'tests.txt')
-
-    def __init__(self, *args, **kwargs):
-        parser = doctest.DocTestParser()
-        fd = open(self.path)
-        doc = fd.read()
-        fd.close()
-        test = parser.get_doctest(doc, globals(), '', self.path, 0)
-        doctest.DocFileCase.__init__(self, test, optionflags=doctest.ELLIPSIS)
-
-
-class TestUnicode(unittest.TestCase):
+class TestUnicode(TestCase):
 
     def test_unicode(self):
-        xml = pq(u("<p>é</p>", 'utf-8'))
+        xml = pq(u("<html><p>é</p></html>", 'utf-8'))
         self.assertEqual(type(xml.html()), text_type)
         if PY3k:
-            self.assertEqual(str(xml), '<p>é</p>')
+            self.assertEqual(str(xml), '<html><p>é</p></html>')
+            self.assertEqual(str(xml('p:contains("é")')), '<p>é</p>')
         else:
-            self.assertEqual(unicode(xml), u("<p>é</p>", 'utf-8'))
-            self.assertEqual(str(xml), '<p>&#233;</p>')
+            self.assertEqual(unicode(xml), u("<html><p>é</p></html>", 'utf-8'))
+            self.assertEqual(str(xml), '<html><p>&#233;</p></html>')
+            self.assertEqual(str(xml(u('p:contains("é")', 'utf8'))),
+                             '<p>&#233;</p>')
+            self.assertEqual(unicode(xml(u('p:contains("é")', 'utf8'))),
+                             u('<p>é</p>', 'utf8'))
 
 
-class TestAttributeCase(unittest.TestCase):
+class TestAttributeCase(TestCase):
 
     def test_xml_upper_element_name(self):
         xml = pq('<X>foo</X>', parser='xml')
@@ -150,7 +66,7 @@ class TestAttributeCase(unittest.TestCase):
         self.assertEqual(len(xml('x')), 1)
 
 
-class TestSelector(unittest.TestCase):
+class TestSelector(TestCase):
     klass = pq
     html = """
            <html>
@@ -288,10 +204,10 @@ class TestSelector(unittest.TestCase):
     def test_on_the_fly_dom_creation(self):
         e = self.klass(self.html)
         assert e('<p>Hello world</p>').text() == 'Hello world'
-        assert e('').text() == None
+        assert e('').text() is None
 
 
-class TestTraversal(unittest.TestCase):
+class TestTraversal(TestCase):
     klass = pq
     html = """
            <html>
@@ -342,12 +258,16 @@ class TestTraversal(unittest.TestCase):
 
     def test_closest(self):
         assert len(self.klass('#node1 span', self.html).closest('body')) == 1
-        assert self.klass('#node2', self.html).closest('.node3').attr('id') \
-                                                                    == 'node2'
+        assert self.klass('#node2',
+                          self.html).closest('.node3').attr('id') == 'node2'
         assert self.klass('.node3', self.html).closest('form') == []
 
 
-class TestOpener(unittest.TestCase):
+class TestOpener(TestCase):
+
+    def test_open_filename(self):
+        doc = pq(filename=path_to_html_file)
+        self.assertEqual(len(doc('p#test').text()), 14)
 
     def test_custom_opener(self):
         def opener(url):
@@ -357,14 +277,14 @@ class TestOpener(unittest.TestCase):
         assert len(doc('.node')) == 1, doc
 
 
-class TestComment(unittest.TestCase):
+class TestComment(TestCase):
 
     def test_comment(self):
         doc = pq('<div><!-- foo --> bar</div>')
         self.assertEqual(doc.text(), 'bar')
 
 
-class TestCallback(unittest.TestCase):
+class TestCallback(TestCase):
     html = """
         <ol>
             <li>Coffee</li>
@@ -384,31 +304,19 @@ class TestCallback(unittest.TestCase):
                                      ['Coffee', 'Tea', 'Milk'])
 
 
-def application(environ, start_response):
-    req = Request(environ)
-    response = Response()
-    if req.method == 'GET':
-        response.body = b('<pre>Yeah !</pre>')
-    else:
-        response.body = b('<a href="/plop">Yeah !</a>')
-    return response(environ, start_response)
-
-
-def secure_application(environ, start_response):
-    if 'REMOTE_USER' not in environ:
-        return exc.HTTPUnauthorized('vomis')(environ, start_response)
-    return application(environ, start_response)
-
-
 class TestAjaxSelector(TestSelector):
     klass = pqa
 
+    def setUp(self):
+        self.s = http.StopableWSGIServer.create(application)
+
     @not_py3k
-    @with_net
     def test_proxy(self):
+        self.s.wait()
+        application_url = self.s.application_url
         e = self.klass([])
-        val = e.get('http://packages.python.org/pyquery/')
-        assert len(val('body')) == 1, (str(val.response), val)
+        val = e.get(application_url)
+        assert len(val('pre')) == 1, (str(val.response), val)
 
     def test_get(self):
         e = self.klass(app=application)
@@ -438,8 +346,11 @@ class TestAjaxSelector(TestSelector):
         val = n.post('/')
         assert len(val('a')) == 1, val
 
+    def tearDown(self):
+        self.s.shutdown()
 
-class TestManipulating(unittest.TestCase):
+
+class TestManipulating(TestCase):
     html = '''
     <div class="portlet">
       <a href="/toto">Test<img src ="myimage" />My link text</a>
@@ -456,7 +367,7 @@ class TestManipulating(unittest.TestCase):
         assert val == ' My link text 2', repr(val)
 
 
-class TestHTMLParser(unittest.TestCase):
+class TestHTMLParser(TestCase):
     xml = "<div>I'm valid XML</div>"
     html = '''
     <div class="portlet">
@@ -503,7 +414,7 @@ class TestHTMLParser(unittest.TestCase):
         assert val == expected, (repr(val), repr(expected))
 
 
-class TestXMLNamespace(unittest.TestCase):
+class TestXMLNamespace(TestCase):
     xml = '''<?xml version="1.0" encoding="UTF-8" ?>
     <foo xmlns:bar="http://example.com/bar">
     <bar:blah>What</bar:blah>
@@ -558,22 +469,37 @@ class TestXMLNamespace(unittest.TestCase):
         self.assertEqual(repr(val), repr(expected))
 
 
-class TestWebScrapping(unittest.TestCase):
+class TestWebScrapping(TestCase):
 
-    @with_net
+    def setUp(self):
+        self.s = http.StopableWSGIServer.create(debug_app)
+        self.s.wait()
+        self.application_url = self.s.application_url.rstrip('/')
+
     def test_get(self):
-        d = pq('http://duckduckgo.com/', {'q': 'foo'},
+        d = pq(self.application_url, {'q': 'foo'},
                method='get')
         print(d)
-        self.assertEqual(d('input[name=q]:last').val(), 'foo')
+        self.assertIn('REQUEST_METHOD: GET', d('p').text())
+        self.assertIn('q=foo', d('p').text())
 
-    @with_net
     def test_post(self):
-        d = pq('http://duckduckgo.com/', {'q': 'foo'},
+        d = pq(self.application_url, {'q': 'foo'},
                method='post')
-        self.assertEqual(d('input[name=q]:last').val(), 'foo')
+        self.assertIn('REQUEST_METHOD: POST', d('p').text())
+        self.assertIn('q=foo', d('p').text())
 
-if __name__ == '__main__':
-    fails, total = unittest.main()
-    if fails == 0:
-        print('OK')
+    def tearDown(self):
+        self.s.shutdown()
+
+
+class TestWebScrappingEncoding(TestCase):
+
+    def test_get(self):
+        if not HAS_REQUEST:
+            return
+        d = pq(u('http://ru.wikipedia.org/wiki/Заглавная_страница', 'utf8'),
+               method='get')
+        print(d)
+        self.assertEqual(d('#n-mainpage a').text(),
+                         u('Заглавная страница', 'utf8'))